combine bare dropdown and dropdown
